+++ zap/src/main/java/org/parosproxy/paros/core/scanner/VariantMultipartFormParameters.java
@@ -110,6 +110,7 @@ public class VariantMultipartFormParameters implements Variant {
                                         + ")?$",
                                 ""); // Strip final boundary
                 if (isFileParam) {
+                    position += 2;
                     extractedParameters.add(
                             new NameValuePair(
                                     NameValuePair.TYPE_MULTIPART_DATA_FILE_PARAM,
@@ -146,16 +147,18 @@ public class VariantMultipartFormParameters implements Variant {
                     LOGGER.debug("Name: " + name + " value: " + valueMatcher.group("value"));
                 }
                 if (isFileParam) {
+                    position -= 2;
                     // Extract the filename
                     Matcher fnValueMatcher = FILENAME_PART_PATTERN.matcher(part);
                     fnValueMatcher.find();
                     String fnValue = fnValueMatcher.group("filename");
                     extractedParameters.add(
+                            extractedParameters.size() - 1,
                             new NameValuePair(
                                     NameValuePair.TYPE_MULTIPART_DATA_FILE_NAME,
                                     name,
                                     fnValue,
-                                    ++position));
+                                    position));
                     int fnStart = offset + part.indexOf(fnValue);
                     int fnEnd = fnStart + fnValue.length();
                     if (LOGGER.isDebugEnabled()) {
@@ -164,6 +167,7 @@ public class VariantMultipartFormParameters implements Variant {
                                         + fnEnd + " Pos: " + position);
                     }
                     multiPartParams.add(
+                            multiPartParams.size() - 1,
                             new MultipartFormParameter(
                                     name,
                                     fnValue,
@@ -176,6 +180,7 @@ public class VariantMultipartFormParameters implements Variant {
                     ctValueMatcher.find();
                     String ctValue = ctValueMatcher.group("contenttype");
                     extractedParameters.add(
+                            extractedParameters.size() - 1,
                             new NameValuePair(
                                     NameValuePair.TYPE_MULTIPART_DATA_FILE_CONTENTTYPE,
                                     name,
@@ -189,6 +194,7 @@ public class VariantMultipartFormParameters implements Variant {
                                         + ctEnd + " Pos: " + position);
                     }
                     multiPartParams.add(
+                            multiPartParams.size() - 1,
                             new MultipartFormParameter(
                                     name,
                                     ctValue,