+++ zap/src/test/java/org/zaproxy/zap/extension/httppanel/view/impl/models/http/request/RequestBodyStringHttpPanelViewModelUnitTest.java
@@ -20,7 +20,9 @@
 package org.zaproxy.zap.extension.httppanel.view.impl.models.http.request;
 
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.verify;
 
+import org.parosproxy.paros.network.HttpMessage;
 import org.parosproxy.paros.network.HttpRequestHeader;
 import org.zaproxy.zap.extension.httppanel.view.impl.models.http.BodyStringHttpPanelViewModelTest;
 import org.zaproxy.zap.network.HttpRequestBody;
@@ -48,5 +50,10 @@ class RequestBodyStringHttpPanelViewModelUnitTest
     protected void prepareMessage() {
         given(message.getRequestHeader()).willReturn(header);
         given(message.getRequestBody()).willReturn(body);
+    }
+
+    @Override
+    protected void verifyBodySet(HttpMessage message, String body) {
+        verify(message).setRequestBody(body);
     }
 }